pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adversarypilot"
version = "1.0.0"
description = "ATLAS-aligned attack planning + orchestration engine for adversarial ML and LLM/agent systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Aviral Srivastava" },
]
keywords = ["adversarial-ml", "llm-security", "red-teaming", "mitre-atlas", "ai-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/aviralsrivastava/AdversaryPilot"
Repository = "https://github.com/aviralsrivastava/AdversaryPilot"
Issues = "https://github.com/aviralsrivastava/AdversaryPilot/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
adversarypilot = "adversarypilot.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/adversarypilot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
